chore: fix build process
As noted in #4, the build task wasn't run when publishing the package to npm. To fix that, we're adding the build-task to the prepack hook, so it will run before a package is created to be uploaded to npm.

Also, since #2 and #3 should fix the tsc, eslint and prettier tests, this also reenabled the full test-suite for the release task.