fix(playground): remove code sharing functions
We decided to remove the save/share functions from the playground since there is a security discussion we should be having first.

The problem we are seeing is that the "Run" functionality essentially allows execution of arbitrary code on the user's machine, which is obviously not ideal (basically, XSS as a feature).

The two options to implement this would be to lock down the execution environment and to make sure users are aware of potential problems, ideally in combination. But since this takes some time to get right, we disable the feature for now.