chore: workflow to deploy website to github pages (#9)
Adds a workflow that builds both the API docs and the
playground application and deploys them to github-pages using the new
actions-based deployment method.

To make this work, Github Pages pages needs to be enabled and configured
to allow deployments from actions. To do that, navigate to Settings >
Code and automation > Pages, then in the section "Build and deployment"
select "Github Actions" as the source. This will allow the
`actions/deploy-pages` action to deploy new versions of the site.

After deployment, the documentation will be available at
https://googlemaps.github.io/js-adv-markers-utils and the playground
will be at https://googlemaps.github.io/js-adv-markers-utils/playground

The Google Maps API key needed for the playground application should be
provided as a secret under the name `GOOGLE_MAPS_API_KEY` (the key can
be limited to the referrer `https://googlemaps.github.io/js-adv-markers-utils/`)

This has already been tested in my fork of the repository at
https://usefulthink.github.io/js-adv-markers-utils/ and
https://usefulthink.github.io/js-adv-markers-utils/playground